All changes as requested
What is the difference between force.com and salesforce.com? -- Changed
Why do we need to write test classes? How to identify if a class is a test class? -- Changed
What are the different ways of deployment in Salesforce? -- Changed
Salesforce should always be capitalized because it is the name of a company, just as Google is always capitalized -- Changed
please don't say force.com developers >>> Salesforce developers -- Changed
Grammer rectification -- Changed
Images added